combined scores
