+++ client/components/bubbles/Fire.js
@@ -1,0 +1,1 @@
+import react from 'export'
