+++ client/components/main/webcam.js
@@ -97,13 +97,13 @@ class PoseNet extends React.Component {
   async componentDidMount() {
     try {
       await this.setupCamera()
+      this.net = await posenet.load()
     } catch (e) {
       throw 'This browser does not support video capture, or this device does not have a camera'
     } finally {
       this.setState({loading: false})
       this.props.sendLoadingState(false)
     }
-    this.net = await posenet.load()
     this.detectPose()
     this.props.onRef(this)
     this.props.danceDancePoseTime()
@@ -198,6 +198,7 @@ class PoseNet extends React.Component {
           )
           // index 10 is rightWrist
           // index 9 is left Wrist
+          //index 13 is right knee
           const {rangeSpectrum, minConfidencePoints} = defaultParameters
           this.setState({
             xMin: this.props.xBubble * (1 - rangeSpectrum),
@@ -209,7 +210,6 @@ class PoseNet extends React.Component {
             yMin2: this.props.yBubble * (1 - rangeSpectrum),
             yMax2: this.props.yBubble * (1 + rangeSpectrum)
           })
-
           if (
             (this.state.xMin < pose.keypoints[10].position.x &&
               pose.keypoints[10].position.x < this.state.xMax &&
@@ -222,16 +222,8 @@ class PoseNet extends React.Component {
               pose.keypoints[13].position.y < this.state.yMax2 &&
               pose.keypoints[3].score > minConfidencePoints)
           ) {
-            counter++
+            counter = counter * 100
             this.props.addScore(counter)
-            console.log(
-              'COUNTER',
-              counter,
-              'STATE',
-              this.state,
-              'POSE',
-              pose.keypoints
-            )
           }
           poses.push(pose)
           break
@@ -320,7 +312,6 @@ class PoseNet extends React.Component {
     const beatsToDisplay = setBeats()
     const beatTime = beatsToDisplay[this.state.counterBeatInterval]
     const level = setLevel()
-    console.log('LEVL', level)
     if (
       this.props.song.destination.context.currentTime - this.state.windowTime >
       beatTime
@@ -345,53 +336,35 @@ class PoseNet extends React.Component {
     this.props.addY(null)
   }
 
-  // render() {
-  //   const loading = this.state.loading
-  //   return (
-  //     <div>
-  //       {loading ? (
-  //         <div>
-  //           <h2>Loading</h2>
-  //           <Loading />
-  //         </div>
-  //       ) : (
-
   render() {
-    const loading = this.state.loading ? (
-      // <h1>HELLO</h1>
-      <ReactLoading
-        type="spinningBubbles"
-        color="white"
-        height="25%"
-        width="25%"
-      />
-    ) : (
-      ''
-    )
+    const loading = this.state.loading ? <h1>LOADING...</h1> : <h1 />
     return (
-      <div id="border">
-        <div className="sidenav" />
-        <div className="score">
-          <Score />
-        </div>
-        <div className="tvlines">
-          <div className="webcam-outer">
-            <video id="notShow" playsInline ref={this.getVideo} />
-            {this.state.time === '' ? (
-              <h2 />
-            ) : (
-              <div id="bubble-container">
-                <Bubble
-                  yBubble={this.props.yBubble}
-                  xBubble={this.props.xBubble}
-                />
-                <Bubble2
-                  yBubble={this.props.yBubble2}
-                  xBubble={this.props.xBubble2}
-                />
-              </div>
-            )}
-            <canvas className="webcam" ref={this.getCanvas} />
+      <div>
+        <div>{loading}</div>
+        <div id="border">
+          <div className="sidenav" />
+          <div className="score">
+            <Score />
+          </div>
+          <div className={this.state.loading ? 'noShow' : 'tvlines'}>
+            <div className="webcam-outer">
+              <video id="notShow" playsInline ref={this.getVideo} />
+              {this.state.time === '' ? (
+                <h2 />
+              ) : (
+                <div id="bubble-container">
+                  <Bubble
+                    yBubble={this.props.yBubble}
+                    xBubble={this.props.xBubble}
+                  />
+                  <Bubble2
+                    yBubble={this.props.yBubble2}
+                    xBubble={this.props.xBubble2}
+                  />
+                </div>
+              )}
+              <canvas className="webcam" ref={this.getCanvas} />
+            </div>
           </div>
         </div>
       </div>
